Create Nicosia Rent Prices
This script mines the data from a popular website that lists apartments for rent. 
The data are parsed in the desired format and they are saved both as a pandas dataframe and as entries in a database. 
The results can then be send by email to the interested parties for additional analysis.